+++ client/src/migrator/validator-utils.js
@@ -1,0 +1,43 @@
+const selectOptions = [
+  { label: 'Story', value: 'Story' },
+  { label: 'Section', value: 'Section' },
+  { label: 'Author', value: 'Author' },
+  { label: 'Entity', value: 'Entity' },
+  { label: 'Tag', value: 'Tag' }
+]
+    
+const validateOptions = [
+  { label: 'Direct text input', value: 'Direct text input' },
+  { label: 'File upload', value: 'File upload'},
+  { label: 'S3 path', value: 'S3 path'}
+]
+
+function createRequest ({value: validateType}, {value: selectType}, userData) {
+  let options = {
+    method: 'POST',
+    headers: {
+      Accept: 'application/json'
+    }
+  }
+  let requestData
+  if(validateType === 'Direct text input' || validateType === 'S3 path') {
+    options.headers['Content-Type'] = 'application/json'
+    requestData = JSON.stringify({
+      type: selectType,
+      data: userData
+    })
+  } else {
+    requestData = new FormData()
+    requestData.append('file', userData)
+    requestData.append('type', selectType)
+  }
+
+  options.body = requestData
+  return options
+}
+
+function parseResult () {
+
+}
+
+export { selectOptions, validateOptions, createRequest, parseResult }
